fix(saplma): cast input to fp32 in SimpleHaluClassifier.forward
Activations from the Zarr/memmap path arrive as fp16, but the classifier
weights are fp32, causing every SAPLMA training run to die in epoch 1
with "mat1 and mat2 must have the same dtype, but got Half and Float".
LinearProbe already handles this with `x = x.float()` (model.py:378);
mirror that pattern here.

This is why the audit shows saplma 0/5 across every dataset+model combo:
the bug silently failed all prior runs in the existing baseline scripts.

Verified end-to-end: sciq_qwen3 seed 1 trains to completion in ~8 min,
test AUROC=0.6174 on layer 22.

+++ activation_research/model.py
@@ -305,10 +305,8 @@ class SimpleHaluClassifier(nn.Module):
         x: (B, L, D)
         returns: (B, 1) sigmoid probability
         """
-        # Take only the last token's activations
+        x = x.float()
         last_token = x[:, -1, :]  # (B, D)
-        
-        # Pass through feed-forward layers
         logits = self.classifier(last_token)
         return torch.sigmoid(logits)
